+++ lib/src/wire/builders/widget_builder.dart
@@ -139,3 +139,5 @@ part 'widget_builders/placeholder_builder.dart';
 part 'widget_builders/ping_error_builder.dart';
 
 part 'widget_builders/center_builder.dart';
+
+part 'widget_builders/choice_chip_builder.dart';